sem2/FastAPI_with_Auth_and_TSP/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
alembic==1.13.2
pydantic-settings==2.5.2
email-validator==2.2.0
python-multipart==0.0.12
bcrypt==4.0.1
pydantic==2.9.2